Move functions into util.go as suggested
